requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2
python-pptx==1.0.2
reportlab==4.2.5
Pillow==10.4.0
pydantic==2.9.2
python-multipart==0.0.12
